+++ packages/base/Avatar/src/Avatar/AvatarWrapper/AvatarWrapper.tsx
@@ -1,19 +1,48 @@
 import type { ReactNode } from 'react'
 import React from 'react'
 import type { BaseProps, SizeType } from '@toptal/picasso-shared'
-import { twMerge } from '@toptal/picasso-tailwind-merge'
+import { twJoin, twMerge } from '@toptal/picasso-tailwind-merge'
+import { Logo } from '@toptal/picasso-logo'
 
-import { classBySizeAndVariant, clipClassBySize } from './styles'
+import { avatarLogoClassesBySize, classBySize, clipClassBySize } from './styles'
 
 export type Size = SizeType<'xxsmall' | 'xsmall' | 'small' | 'medium' | 'large'>
-
-export type Variant = 'square' | 'portrait' | 'landscape'
 
 export interface Props extends BaseProps {
   children: ReactNode
   size: Size
-  variant: Variant
+  showEmblem?: boolean
 }
+
+const showLogo = (size: Size, showEmblem?: boolean): boolean => {
+  const isTooSmall = ['xsmall', 'xxsmall'].includes(size)
+
+  if (isTooSmall || !showEmblem) {
+    return false
+  }
+
+  return true
+}
+
+const AvatarLogo = ({ size }: Pick<Props, 'size'>) => (
+  <div
+    className={twJoin(
+      'flex absolute bottom-0',
+      avatarLogoClassesBySize.root[size]
+    )}
+    role='img'
+    aria-label='photo placeholder'
+  >
+    <Logo
+      emblem
+      variant='blue'
+      // Necessary to decrease the default minWeight and minHeight so that the new width and height could apply
+      // eslint-disable-next-line no-inline-styles/no-inline-styles
+      style={{ minWidth: '1px', minHeight: '1px' }}
+      className={avatarLogoClassesBySize.logo[size]}
+    />
+  </div>
+)
 
 const AvatarWrapper = (props: Props) => {
   const {
@@ -22,21 +51,30 @@ const AvatarWrapper = (props: Props) => {
     style,
     'data-testid': dataTestId,
     size,
-    variant,
+    showEmblem = false,
   } = props
 
   return (
     <div
-      style={style}
-      data-testid={dataTestId}
-      className={twMerge(
-        'relative bg-gray-500 text-[1rem] shrink-0 grow-0',
-        classBySizeAndVariant[size][variant],
-        clipClassBySize[size],
-        className
+      className={twJoin(
+        `flex relative`,
+        showLogo(size, showEmblem) && 'bg-white'
       )}
     >
-      {children}
+      <div
+        style={style}
+        data-testid={dataTestId}
+        className={twMerge(
+          'relative bg-gray-500 text-[1rem] shrink-0 grow-0',
+          classBySize[size],
+          clipClassBySize[size],
+          className
+        )}
+      >
+        {children}
+      </div>
+
+      {showLogo(size, showEmblem) && <AvatarLogo size={size} />}
     </div>
   )
 }