+++ packages/base/EmptyState/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-empty-state
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`cd00880`](https://github.com/toptal/picasso/commit/cd008802e9701bb6574b3354f9d6e95c6ffc61ed)]:
+  - @toptal/picasso-icons@1.5.0
+
 ## 1.0.10
 
 ### Patch Changes