+++ packages/base/Radio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @toptal/picasso-radio
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-form@3.0.7
+  - @toptal/picasso-grid@3.0.6
+
 ## 2.0.6
 
 ### Patch Changes