+++ packages/base/Switch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-switch
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-form@3.0.7
+
 ## 2.0.2
 
 ### Patch Changes